requirements.txt
fastapi>=0.100.0
uvicorn>=0.22.0
pydantic>=2.0.0
sqlalchemy>=2.0.23
asyncpg>=0.29.0
psycopg2-binary>=2.9.9
alembic>=1.12.1
python-dotenv>=1.0.0
pydantic-settings>=2.0.3 